Change code structure.

+++ GitCommit.sh
@@ -23,30 +23,39 @@ while read Line; do
     export ${Var[0]}=${Var[1]}
 done < local.conf
 
+# Check if git repository directory exists
+ls ${RepoPath} > /dev/null 2> /dev/null
 
-# Check directory
-ls ${CommitPath} > /dev/null 2> /dev/null
-if [ $? != 0 ]; then
-    mkdir ${CommitPath} > /dev/null
+if [ $? == 0 ]; then
+
+    # Check directory
+    ls ${CommitPath} > /dev/null 2> /dev/null
     if [ $? != 0 ]; then
-        exit 1
+
+        mkdir ${CommitPath} > /dev/null
+        if [ $? != 0 ]; then
+            exit 1
+        fi
+
+    else
+        rm ${CommitPath}"/*" > /dev/null 2> /dev/null
     fi
+
+
+    # Find all Git Repositories
+    find ${RepoPath} -name '*.git' | while read Repo; do
+
+        cd "$Repo"
+        echo "Processing Git Repository -> '$Repo'"
+
+        RepoName=${Repo/${RepoPath}/""}
+        CommitFile=${CommitPath}"/"$RepoName"CommitLog.txt"
+
+        # Read Commits from the Git Repositories
+        echo -e "\n\n\tGit Repository: $RepoName\n" >> ${CommitFile}
+
+        git log --pretty=format:"%h - %an, %ad : %s" --graph >> ${CommitFile}
+
+    done
+
 fi
-
-rm ${CommitPath}"/*" > /dev/null 2> /dev/null
-
-# Find all Git Repositories
-find ${RepoPath} -name '*.git' | while read Repo; do
-
-    cd "$Repo"
-    echo "Processing Git Repository -> '$Repo'"
-
-    RepoName=${Repo/${RepoPath}/""}
-    CommitFile=${CommitPath}"/"$RepoName"CommitLog.txt"
-
-    # Read Commits from the Git Repositories
-    echo -e "\n\n\tGit Repository: $RepoName\n" >> ${CommitFile}
-
-    git log --pretty=format:"%h - %an, %ad : %s" --graph >> ${CommitFile}
-
-done
